+++ main.js
@@ -29,24 +29,24 @@
  * @see https://encoding.spec.whatwg.org/#utf-8-encoder
  */
 
-import utf8BufferSize from 'utf8-buffer-size';
+/** @module utf8-buffer */
 
 /**
  * Read a string of UTF-8 characters from a byte buffer.
  * Invalid characters are replaced with 'REPLACEMENT CHARACTER' (U+FFFD).
  * @see https://encoding.spec.whatwg.org/#the-encoding
  * @see https://stackoverflow.com/a/34926911
- * @param {!Uint8Array|!Array<!number>} buffer A byte buffer.
- * @param {number=} index The index to read.
- * @param {?number=} len The number of bytes to read.
- *    If len is undefined will read until the end of the buffer.
+ * @param {!Uint8Array|!Array<number>} buffer A byte buffer.
+ * @param {number=} start The buffer index to start reading.
+ * @param {?number=} end The buffer index to stop reading.
+ *    If end is null will read until the end of the buffer.
  * @return {string}
  */
-export function unpack(buffer, index=0, len=undefined) {
-  len = len !== undefined ? index + len : buffer.length;
+export function unpack(buffer, start=0, end=null) {
+  end = end !== null ? end + 1 : buffer.length;
   /** @type {string} */
   let str = "";
-  while(index < len) {
+  for(let index = start; index < end;) {
     /** @type {number} */
     let lowerBoundary = 0x80;
     /** @type {number} */
@@ -106,21 +106,20 @@ export function unpack(buffer, index=0, len=undefined) {
 }
 
 /**
- * Write a string of UTF-8 characters as a byte buffer.
+ * Write a string of UTF-8 characters to a byte buffer.
  * @see https://encoding.spec.whatwg.org/#utf-8-encoder
  * @param {string} str The string to pack.
- * @return {!Uint8Array} The packed string.
+ * @param {!Uint8Array|!Array<number>} buffer The buffer to pack the string to.
+ * @param {number=} index The buffer index to start writing.
+ * @return {number} The next index to write in the buffer.
  */
-export function pack(str) {
-  /** @type {!Uint8Array} */
-  let bytes = new Uint8Array(utf8BufferSize(str));
-  let bufferIndex = 0;
+export function pack(str, buffer, index=0) {
   for (let i = 0, len = str.length; i < len; i++) {
     /** @type {number} */
     let codePoint = str.codePointAt(i);
     if (codePoint < 128) {
-      bytes[bufferIndex] = codePoint;
-      bufferIndex++;
+      buffer[index] = codePoint;
+      index++;
     } else {
       /** @type {number} */
       let count = 0;
@@ -137,14 +136,14 @@ export function pack(str) {
         offset = 0xF0;
         i++;
       }
-      bytes[bufferIndex] = (codePoint >> (6 * count)) + offset;
-      bufferIndex++;
+      buffer[index] = (codePoint >> (6 * count)) + offset;
+      index++;
       while (count > 0) {
-        bytes[bufferIndex] = 0x80 | (codePoint >> (6 * (count - 1)) & 0x3F);
-        bufferIndex++;
+        buffer[index] = 0x80 | (codePoint >> (6 * (count - 1)) & 0x3F);
+        index++;
         count--;
       }
     }
   }
-  return bytes;
+  return index++;
 }