+++ rollup.config.js
@@ -8,8 +8,6 @@
  */
 
 import closure from 'rollup-plugin-closure-compiler-js';
-import resolve from 'rollup-plugin-node-resolve';
-import commonjs from 'rollup-plugin-commonjs';
 import fs from 'fs';
 
 // Externs
@@ -21,12 +19,12 @@ const license = '/*!\n'+
   ' * Copyright (c) 2018 Rafael da Silva Rocha.\n' +
   ' */\n';
 
-// GCC wrapper
-const outputWrapper = license + '"use strict";if(typeof exports!=="undefined"){var window={};}'+
-  '%output%' +
+// GCC UMD wrapper
+const outputWrapper = '%output%' +
   'var module=module||{};module.exports=exports;' +
   'var define=define||function(){};' +
-  'define(["exports"],function(e){return module.exports;});'
+  'define(["exports"],function(e){return module.exports;});' +
+  'var utf8Buffer=exports;'
 
 export default [
   // ES6 bundle
@@ -37,36 +35,28 @@ export default [
         file: 'dist/utf8-buffer.js',
         format: 'es'
       },
-    ],
-    plugins: [
-      resolve(),
-      commonjs()
     ]
   },
-  // ES5 UMD
+  // UMD, minified
   {
-    input: 'main.js',
+    input: 'dist/utf8-buffer.js',
     output: [
       {
         file: 'dist/utf8-buffer.umd.js',
-        name: 'utf8Buffer',
         format: 'cjs',
         strict: false,
-        banner: 'var exports=exports||{};' +
-        'if(window){window["utf8Buffer"]=exports;}'
+        banner: 'var exports=exports||{};'
       }
     ],
     plugins: [
-      resolve(),
-      commonjs(),
       closure({
         languageIn: 'ECMASCRIPT6',
         languageOut: 'ECMASCRIPT5',
         compilationLevel: 'ADVANCED',
         warningLevel: 'VERBOSE',
-        outputWrapper: outputWrapper,
+        outputWrapper: license + outputWrapper,
         externs: [{src: externsFile + 'exports={};'}]
-      }),
+      })
     ]
   },
 ];