+++ test/src/UTF-8-with-arrays.js
@@ -61,9 +61,9 @@ describe('unpack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 0xEF,0xBB,0xBF,
-                240,175,167,159]),
+                240,175,167,159],
             buffer);
     });
 });
@@ -74,10 +74,10 @@ describe('unpack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 240,175,167,159,
                 0xEF,0xBB,0xBF,
-                240,175,167,159]),
+                240,175,167,159],
             buffer);
     });
 });
@@ -88,7 +88,7 @@ describe('pack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([240,175,167,159]),
+            [240,175,167,159],
             buffer);
     });
     it('pack 輸輸 as a byte array', function() {
@@ -96,9 +96,9 @@ describe('pack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 240,175,167,159,
-                240,175,167,159]),
+                240,175,167,159],
             buffer);
     });
     it('pack 輸輸笠߹~$ as a byte array', function() {
@@ -106,13 +106,13 @@ describe('pack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 240,175,167,159,
                 240,175,167,159,
                 239, 167, 184,
                 223, 185,
                 126,
-                36]),
+                36],
             buffer);
     });
     it('pack 輸輸笠߹~$輸 as a byte array', function() {
@@ -120,14 +120,14 @@ describe('pack UTF-8 strings, 4 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 240,175,167,159,
                 240,175,167,159,
                 239, 167, 184,
                 223, 185,
                 126,
                 36,
-                240,175,167,159]),
+                240,175,167,159],
             buffer);
     });
 });
@@ -138,7 +138,7 @@ describe('pack UTF-8 strings, 3 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([239, 167, 184]),
+            [239, 167, 184],
             buffer);
     });
     it('pack 笠笠 as a byte array', function() {
@@ -146,9 +146,9 @@ describe('pack UTF-8 strings, 3 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 239, 167, 184,
-                239, 167, 184]),
+                239, 167, 184],
             buffer);
     });
     it('pack 笠笠߹~$ as a byte array', function() {
@@ -156,12 +156,12 @@ describe('pack UTF-8 strings, 3 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 239, 167, 184,
                 239, 167, 184,
                 223, 185,
                 126,
-                36]),
+                36],
             buffer);
     });
     it('pack 笠笠߹~$笠 as a byte array', function() {
@@ -169,13 +169,13 @@ describe('pack UTF-8 strings, 3 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([
+            [
                 239, 167, 184,
                 239, 167, 184,
                 223, 185,
                 126,
                 36,
-                239, 167, 184]),
+                239, 167, 184],
             buffer);
     });
 });
@@ -185,7 +185,7 @@ describe('pack UTF-8 strings, 2 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([223, 185]),
+            [223, 185],
             buffer);
     });
     it('pack ߹߹ as a byte array', function() {
@@ -193,7 +193,7 @@ describe('pack UTF-8 strings, 2 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([223, 185, 223, 185]),
+            [223, 185, 223, 185],
             buffer);
     });
     it('pack ߹߹~$ as a byte array', function() {
@@ -201,7 +201,7 @@ describe('pack UTF-8 strings, 2 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([223, 185, 223, 185, 126, 36]),
+            [223, 185, 223, 185, 126, 36],
             buffer);
     });
     it('pack ߹߹~$߹ as a byte array', function() {
@@ -209,7 +209,7 @@ describe('pack UTF-8 strings, 2 bytes', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([223, 185, 223, 185, 126, 36, 223, 185]),
+            [223, 185, 223, 185, 126, 36, 223, 185],
             buffer);
     });
 });
@@ -219,7 +219,7 @@ describe('pack UTF-8 strings, 1 byte', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([126]),
+            [126],
             buffer);
     });
     it('pack ~~ as a byte array', function() {
@@ -227,7 +227,7 @@ describe('pack UTF-8 strings, 1 byte', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
        assert.deepEqual(
-            new Uint8Array([126, 126]),
+            [126, 126],
             buffer);
     });
     it('pack ~~$ as a byte array', function() {
@@ -235,7 +235,7 @@ describe('pack UTF-8 strings, 1 byte', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([126, 126, 36]),
+            [126, 126, 36],
             buffer);
     });
     it('pack ~~$~ as a byte array', function() {
@@ -243,7 +243,7 @@ describe('pack UTF-8 strings, 1 byte', function() {
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
         assert.deepEqual(
-            new Uint8Array([126, 126, 36, 126]),
+            [126, 126, 36, 126],
             buffer);
     });
 });
@@ -252,13 +252,13 @@ describe('pack ASCII strings', function() {
         var chars = '$';
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
-        assert.deepEqual(new Uint8Array([36]), buffer);
+        assert.deepEqual([36], buffer);
     });
     it('pack $$ as a byte array', function() {
         var chars = '$$';
         var buffer = [];
         utf8Buffer.pack(chars, buffer);
-        assert.deepEqual(new Uint8Array([36, 36]), buffer);
+        assert.deepEqual([36, 36], buffer);
     });
 });